requirements.txt
phidata
openai
packaging
